# docs/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SSAC Backend API",
...
    "contact": {
      "name": "SSAC Backend Team",
      "email": "ssac-backend@ssac.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users/me/logout": {
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "\uc804\uccb4 \ub514\ubc14\uc774\uc2a4 \ub85c\uadf8\uc544\uc6c3",
        "description": "[\ud638\ucd9c \ud654\uba74] \ub9c8\uc774\ud398\uc774\uc9c0 > \uc124\uc815 > \ubcf4\uc548 \uad00\ub9ac > \ubaa8\ub4e0 \uae30\uae30\uc5d0\uc11c \ub85c\uadf8\uc544\uc6c3 \ud074\ub9ad \uc2dc.\n[\uad8c\ud55c \uc870\uac74] \ub85c\uadf8\uc778 \ud68c\uc6d0 \uc804\uc6a9 (USER, ADMIN).\n[\ud2b9\uc774 \ub3d9\uc791] \ud574\ub2f9 \uacc4\uc815\uc73c\ub85c \ubc1c\uae09\ub41c \ubaa8\ub4e0 Refresh Token\uc744 \ubb34\ud6a8\ud654\ud558\uace0 \uad00\ub828 \ucfe0\ud0a4\ub97c \ubaa8\ub450 \uc0ad\uc81c\ud55c\ub2e4.\n",
        "operationId": "logoutAll",
        "responses": {
          "403": {
            "description": "loginRequired: true (\ud68c\uc6d0 \ub85c\uadf8\uc778 \ud544\uc694)"
          },
          "204": {
            "description": "\uc804\uccb4 \ub514\ubc14\uc774\uc2a4 \ub85c\uadf8\uc544\uc6c3 \uc131\uacf5"
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/users/nickname": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "\ub2c9\ub124\uc784 \uc218\uc815 (\ub2e8\ucd95 \uacbd\ub85c)",
        "description": "[\ud638\ucd9c \ud654\uba74] \ub9c8\uc774\ud398\uc774\uc9c0 > \ub2c9\ub124\uc784 \uc218\uc815.\n[\uad8c\ud55c \uc870\uac74] \ub85c\uadf8\uc778 \ud68c\uc6d0 \uc804\uc6a9 (USER, ADMIN).\n",
        "operationId": "updateNickname",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNicknameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "\ub2c9\ub124\uc784 \uc218\uc815 \uc131\uacf5"
          },
          "409": {
            "description": "\uc774\ubbf8 \uc0ac\uc6a9 \uc911\uc778 \ub2c9\ub124\uc784"
          },
          "400": {
            "description": "\ub2c9\ub124\uc784 \uc720\ud6a8\uc131 \uac80\uc0ac \uc2e4\ud328"
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/users/me/nickname": {
      "patch": {
        "tags": [
          "Profile"
        ],
        "summary": "\ub2c9\ub124\uc784 \uc218\uc815",
        "description": "[\ud638\ucd9c \ud654\uba74] \ub9c8\uc774\ud398\uc774\uc9c0 > \ud504\ub85c\ud544 \uc218\uc815 \uc139\uc158\uc5d0\uc11c \ub2c9\ub124\uc784 \uc218\uc815 \uc2dc \ud638\ucd9c.\n[\uad8c\ud55c \uc870\uac74] \ub85c\uadf8\uc778 \ud68c\uc6d0 \uc804\uc6a9 (USER, ADMIN).\n[\ud2b9\uc774 \ub3d9\uc791] \ub2c9\ub124\uc784 \uc911\ubcf5 \uc2dc 409 \ubc18\ud658. \ub2c9\ub124\uc784 \uc815\ucc45: 2~20\uc790, \ud2b9\uc218\ubb38\uc790 \uc81c\ud55c.\n",
        "operationId": "updateNickname_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNicknameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "\ub2c9\ub124\uc784 \uc218\uc815 \uc131\uacf5"
          },
          "403": {
            "description": "loginRequired: true (\ud68c\uc6d0 \ub85c\uadf8\uc778 \ud544\uc694)"
          },
          "404": {
            "description": "\uc0ac\uc6a9\uc790\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc74c"
          },
          "409": {
            "description": "\uc774\ubbf8 \uc0ac\uc6a9 \uc911\uc778 \ub2c9\ub124\uc784"
          },
          "400": {
            "description": "\ub2c9\ub124\uc784 \uc720\ud6a8\uc131 \uac80\uc0ac \uc2e4\ud328"
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/users/mypage": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "\ub9c8\uc774\ud398\uc774\uc9c0 \ud504\ub85c\ud544 \uc870\ud68c",
        "description": "[\ud638\ucd9c \ud654\uba74] \ub9c8\uc774\ud398\uc774\uc9c0 \uc9c4\uc785 \uc2dc \ud638\ucd9c.\n[\uad8c\ud55c \uc870\uac74] \ub85c\uadf8\uc778 \ud68c\uc6d0 \uc804\uc6a9 (USER, ADMIN).\n[\ud2b9\uc774 \ub3d9\uc791] \ub204\uc801 \ud1b5\uacc4, \uc5f0\uc18d \ud559\uc2b5\uc77c, \uad00\uc2ec \ub3c4\uba54\uc778\uc744 \ud3ec\ud568\ud55c \ud504\ub85c\ud544 \ubc18\ud658.\n",
        "operationId": "getMyPage",
        "responses": {
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "200": {
            "description": "\ub9c8\uc774\ud398\uc774\uc9c0 \uc870\ud68c \uc131\uacf5",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMyPageResponse"
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "\ub0b4 \ud504\ub85c\ud544 \uc870\ud68c",
        "description": "[\ud638\ucd9c \ud654\uba74] \ub9c8\uc774\ud398\uc774\uc9c0 \uc9c4\uc785 \uc2dc \ud638\ucd9c.\n[\uad8c\ud55c \uc870\uac74] \ub85c\uadf8\uc778 \ud68c\uc6d0 \uc804\uc6a9 (USER, ADMIN). \ube44\ud68c\uc6d0(GUEST) \uc811\uadfc \ubd88\uac00.\n[\ud2b9\uc774 \ub3d9\uc791] \ud1a0\ud070\uc758 \uc774\uba54\uc77c\ub85c \uc0ac\uc6a9\uc790\ub97c \uc2dd\ubcc4\ud558\uba70 \ud0c0\uc778 \ud504\ub85c\ud544 \uc870\ud68c \ubd88\uac00.\n",
        "operationId": "getProfile",
        "responses": {
          "403": {
            "description": "loginRequired: true (\ud68c\uc6d0 \ub85c\uadf8\uc778 \ud544\uc694)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMyPageResponse"
                }
              }
            }
          },
          "404": {
            "description": "\uc0ac\uc6a9\uc790\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc74c",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMyPageResponse"
                }
              }
            }
          },
          "200": {
            "description": "\ud504\ub85c\ud544 \uc870\ud68c \uc131\uacf5",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMyPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "\ud68c\uc6d0 \ud0c8\ud1f4",
        "description": "[\ud638\ucd9c \ud654\uba74] \ub9c8\uc774\ud398\uc774\uc9c0 > \ud68c\uc6d0 \ud0c8\ud1f4.\n[\uad8c\ud55c \uc870\uac74] \ub85c\uadf8\uc778 \ud68c\uc6d0 \uc804\uc6a9 (USER, ADMIN).\n[\ud2b9\uc774 \ub3d9\uc791] Soft delete + \uac1c\uc778\uc815\ubcf4 \uc775\uba85\ud654 + \ucfe0\ud0a4 \uc0ad\uc81c.\n",
        "operationId": "withdraw",
        "responses": {
          "204": {
            "description": "\ud68c\uc6d0 \ud0c8\ud1f4 \uc131\uacf5"
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseError": {
        "properties": {
          "success": {
            "example": false
          },
          "data": {
            "example": null
          },
          "message": {
            "example": "\uc0ac\uc6a9\uc790\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4."
          }
        }
      },
...
      "UpdateNicknameRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "\ubcc0\uacbd\ud560 \ub2c9\ub124\uc784",
            "example": "\uc0c8\ub2c9\ub124\uc784123",
            "maxLength": 20,
            "minLength": 2,
            "pattern": "^[a-zA-Z0-9\uac00-\ud7a3_-]+$"
          }
        },
        "required": [
          "nickname"
        ]
      },
...
      "ApiResponseMyPageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/MyPageResponse"
          },
          "message": {
            "type": "string"
          },
          "loginRequired": {
            "type": "boolean"
          }
        }
      },
      "MyPageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "userTypeLabel": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "levelLabel": {
            "type": "string"
          },
          "levelEmoji": {
            "type": "string"
          },
          "onboardingCompleted": {
            "type": "boolean"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/StatsDto"
          },
          "provider": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StatsDto": {
        "type": "object",
        "properties": {
          "totalContentsCompleted": {
            "type": "integer",
            "format": "int64"
          },
          "totalQuizCompleted": {
            "type": "integer",
            "format": "int64"
          },
          "correctRate": {
            "type": "integer",
            "format": "int32"
          },
          "continuousLearningDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "\ub85c\uadf8\uc778 \ud6c4 \ubc1c\uae09\ubc1b\uc740 JWT \ud1a0\ud070. \ud615\uc2dd: Bearer <token>",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}